First pass textfield getStyles
