+++ packages/office-ui-fabric-react/src/components/DatePicker/DatePicker.tsx
@@ -10,7 +10,7 @@ import {
 import { FirstWeekOfYear } from '../../utilities/dateValues/DateValues';
 import { Callout } from '../../Callout';
 import { DirectionalHint } from '../../common/DirectionalHint';
-import { TextField } from '../../TextField';
+import { TextField, TextFieldBase } from '../../TextField';
 import {
   autobind,
   BaseComponent,
@@ -123,7 +123,7 @@ export class DatePicker extends BaseComponent<IDatePickerProps, IDatePickerState
   private _root: HTMLElement;
   private _calendar: Calendar;
   private _datepicker: HTMLDivElement;
-  private _textField: TextField;
+  private _textField: TextFieldBase;
   private _preventFocusOpeningPicker: boolean;
   private _focusOnSelectedDateOnUpdate: boolean;
 
@@ -216,7 +216,7 @@ export class DatePicker extends BaseComponent<IDatePickerProps, IDatePickerState
             } }
             readOnly={ !allowTextInput }
             value={ formattedDate }
-            ref={ this._resolveRef('_textField') }
+            componentRef={ this._resolveRef('_textField') }
             role={ allowTextInput ? 'combobox' : 'menu' }
           />
         </div>
