+++ packages/office-ui-fabric-react/src/components/TextField/index.ts
@@ -1,2 +1,3 @@
 export * from './TextField';
+export * from './TextField.base';
 export * from './TextField.types';
